// types/enums.ts
export enum Gender {
  MALE = 'Male',
  FEMALE = 'Female'
}

export enum Genre {
  POP = 'Pop',
  HIP_HOP = 'Hip-Hop',
  RNB = 'R&B',
  EDM = 'EDM',
  KPOP = 'K-Pop',
  LATIN = 'Latin',
  ROCK = 'Rock',
  AFROBEATS = 'Afrobeats'
}

export enum View {
  CREATE_ARTIST = 'CREATE_ARTIST',
  DASHBOARD = 'DASHBOARD',
  STUDIO = 'STUDIO',
  WORLD_MAP = 'WORLD_MAP',
  TRENDS = 'TRENDS',
  CHARTS = 'CHARTS',
  SPOTIFY = 'SPOTIFY',
  CHART_HISTORY = 'CHART_HISTORY',
  TOUR = 'TOUR',
  YOUTUBE = 'YOUTUBE',
  SOCIAL = 'SOCIAL',
  HALL_OF_FAME = 'HALL_OF_FAME',
  SHOP = 'SHOP',
  GRAMMYS = 'GRAMMYS'
}